refactor the file structure of sql converters
